SnapdragonMusic: exit when quick click in now playing
Music will exit when click quickly in the now playing.

Caused by the bitmap recycled by gc.
We will check the bitmap is recycled or not, and set the
default bitmap when get the empty artwork, furthermore,
we use a cache to hold the bitmap in the memory for a longer
life cycle.

Change-Id: Ifea30e844d7bfebcaf2d7f9494188d2438dc8719
CRs-Fixed: 2437489